Be more intentional about float vs double types.
Try to use double outside of interactions with processing.

+++ src/diffeq/BaseFirstOrderODE.java
@@ -1,10 +1,10 @@
 package diffeq;
 
-import static processing.core.PApplet.atan2;
+import static java.lang.Math.atan2;
 
 public abstract class BaseFirstOrderODE implements SlopeFunction, FirstOrderODE {
     @Override
-    public float getSlope(float x, float y) {
+    public double getSlope(double x, double y) {
         return atan2(dy_over_dt(x, y), dx_over_dt(x, y));
     }
 }
