+++ src/diffeq/FirstOrderODE.java
@@ -1,7 +1,7 @@
 package diffeq;
 
 public interface FirstOrderODE {
-    public float dx_over_dt(float x, float y);
+    public double dx_over_dt(double x, double y);
 
-    public float dy_over_dt(float x, float y);
+    public double dy_over_dt(double x, double y);
 }
