+++ src/diffeq/SlopeFunction.java
@@ -1,5 +1,5 @@
 package diffeq;
 
 public interface SlopeFunction {
-    public float getSlope(float x, float y);
+    public double getSlope(double x, double y);
 }
