+++ src/diffeq/examples/CircleODE.java
@@ -4,12 +4,12 @@ import diffeq.BaseFirstOrderODE;
 
 public class CircleODE extends BaseFirstOrderODE {
     @Override
-    public float dx_over_dt(float x, float y) {
+    public double dx_over_dt(double x, double y) {
         return y;
     }
 
     @Override
-    public float dy_over_dt(float x, float y) {
+    public double dy_over_dt(double x, double y) {
         return -x;
     }
 }
